+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.2b1
 
+* Add retry on 500 server errors from Zaptec cloud, #90
 * Remove large and noisy charger states containg test or production data.
   Helps to reduce the entity attribute <16kb. #88
 * Ignore updates with chargerid 000000...000, #84